Improve backoffice code organization
