+++ apps/backoffice/src/components/catalog-queue/realtime.tsx
@@ -11,89 +11,19 @@ import type {
 import { parseCatalogMaintenanceQueueSnapshotMessage } from '../../lib/catalogMaintenanceQueueLive';
 import { formatLiveSyncTime } from '../liveRefreshTime';
 import { SimplePaginationControls } from '../shared';
+import {
+  buildQueueHref,
+  getLastQueueEvent,
+  getQueueHealth,
+  getQueueJobLinks,
+  getQueueStateClass,
+  getQueueStateCount,
+  QUEUE_STATES,
+  queueRealtimeCopy,
+  STATE_LABELS,
+} from './helpers';
 
 type RealtimeStatus = 'connecting' | 'connected' | 'error';
-
-const QUEUE_STATES = [
-  'waiting',
-  'active',
-  'delayed',
-  'failed',
-  'completed',
-] as const satisfies readonly CatalogMaintenanceQueueJobState[];
-
-const STATE_LABELS: Record<CatalogMaintenanceQueueJobState, string> = {
-  active: 'Active',
-  completed: 'Completed',
-  delayed: 'Scheduled',
-  failed: 'Failed',
-  waiting: 'Waiting',
-};
-
-function getQueueStateCount(
-  jobPage: CatalogMaintenanceQueueJobPage,
-  state: CatalogMaintenanceQueueJobState,
-): number {
-  if (state === 'waiting') return jobPage.counts.waiting;
-  return jobPage.counts[state];
-}
-
-function getQueueStateClass(state: CatalogMaintenanceQueueJobState): string {
-  if (state === 'failed') return 'failed';
-  if (state === 'completed') return 'completed';
-  if (state === 'active') return 'active';
-  return 'queued';
-}
-
-function buildQueueHref({
-  page,
-  pageSize,
-  state,
-}: {
-  page: number;
-  pageSize: number;
-  state: CatalogMaintenanceQueueJobState;
-}) {
-  const params = new URLSearchParams();
-  params.set('state', state);
-  params.set('page', String(page));
-  params.set('pageSize', String(pageSize));
-  return `/queue?${params.toString()}`;
-}
-
-function getQueueHealth(jobPage: CatalogMaintenanceQueueJobPage): {
-  state: 'healthy' | 'warning' | 'active' | 'unavailable';
-  title: string;
-  copy: string;
-} {
-  if (!jobPage.available) {
-    return {
-      state: 'unavailable',
-      title: 'Queue connection unavailable',
-      copy: 'Set REDIS_URL and check backoffice logs before enqueueing repairs.',
-    };
-  }
-  if (jobPage.counts.failed > 0) {
-    return {
-      state: 'warning',
-      title: 'Failed catalog jobs need review',
-      copy: 'Open failed jobs first, then follow the movie or batch link before retrying in Bull Board.',
-    };
-  }
-  if (jobPage.openJobs > 0) {
-    return {
-      state: 'active',
-      title: 'Catalog maintenance has open work',
-      copy: 'This view updates as workers move jobs through waiting, active, and scheduled states.',
-    };
-  }
-
-  return {
-    state: 'healthy',
-    title: 'Catalog maintenance queue is clear',
-    copy: 'No open catalog-maintenance jobs are waiting, active, or scheduled.',
-  };
-}
 
 function QueueStateTabs({ jobPage }: { jobPage: CatalogMaintenanceQueueJobPage }) {
   return (
@@ -238,20 +168,7 @@ function QueueJobDetails({ job }: { job: CatalogMaintenanceQueueJobSummary }) {
 }
 
 function QueueJobLinks({ job }: { job: CatalogMaintenanceQueueJobSummary }) {
-  const links: Array<{ href: string; label: string }> = [];
-
-  if (job.movieId) {
-    links.push({
-      href: `/movies/${encodeURIComponent(job.movieId)}`,
-      label: `Movie ${job.movieId}`,
-    });
-  }
-  if (job.repairBatchId) {
-    links.push({
-      href: `/repair-batches/${encodeURIComponent(job.repairBatchId)}`,
-      label: `Batch ${job.repairBatchId}`,
-    });
-  }
+  const links = getQueueJobLinks(job);
 
   if (links.length === 0) return <span className="muted">-</span>;
 
@@ -264,12 +181,6 @@ function QueueJobLinks({ job }: { job: CatalogMaintenanceQueueJobSummary }) {
       ))}
     </div>
   );
-}
-
-function getLastQueueEvent(job: CatalogMaintenanceQueueJobSummary): string {
-  if (job.finishedAt) return `Finished ${formatLiveSyncTime(job.finishedAt)}`;
-  if (job.processedAt) return `Started ${formatLiveSyncTime(job.processedAt)}`;
-  return job.createdAt ? `Created ${formatLiveSyncTime(job.createdAt)}` : 'Created unknown';
 }
 
 function QueueJobRows({
@@ -330,12 +241,7 @@ function QueueRealtimeStatus({
   lastEventAt: string | null;
   status: RealtimeStatus;
 }) {
-  const copy =
-    status === 'connected'
-      ? 'Queue updates are live'
-      : status === 'connecting'
-        ? 'Connecting to live updates'
-        : 'Live updates are reconnecting';
+  const copy = queueRealtimeCopy(status);
   const lastEvent = lastEventAt ? formatLiveSyncTime(lastEventAt) : null;
 
   return (
