+++ apps/backoffice/src/components/repair-batches/filterLink.tsx
@@ -1,0 +1,15 @@
+export function FilterLink({
+  active,
+  href,
+  label,
+}: {
+  active: boolean;
+  href: string;
+  label: string;
+}) {
+  return (
+    <a className={`filter-chip ${active ? 'active' : ''}`} href={href}>
+      {label}
+    </a>
+  );
+}
